requirements.txt
openai>=1.0.0,<2.0.0
deepgram-sdk>=3.1.0,<4.0.0
elevenlabs>=1.0.0,<2.0.0
numpy>=1.26,<2.0
sounddevice>=0.4.6,<0.5
python-dotenv>=1.0,<2.0
sniffio>=1.3.0
